Add check for MANIFEST file within uploaded archive. Fixes to UI JS.

+++ csp-apps/configuration/conf-server/src/main/java/com/intrasoft/csp/conf/server/controller/DataController.java
@@ -21,6 +21,7 @@ import com.intrasoft.csp.conf.server.repository.*;
 import com.intrasoft.csp.conf.server.utils.FileHelper;
 import com.intrasoft.csp.conf.server.utils.JodaConverter;
 import com.intrasoft.csp.conf.server.utils.VersionParser;
+import org.apache.commons.io.FileUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -31,6 +32,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.io.File;
 import java.io.IOException;
 import java.nio.file.NoSuchFileException;
 import java.security.NoSuchAlgorithmException;
@@ -591,6 +593,18 @@ public class DataController implements DataContextUrl, PagesContextUrl {
         try {
             String hash = FileHelper.saveUploadedFile(fileTemp, fileRepository, uploadFile, digestAlgorithm);
 
+            //check for manifest.json within archive
+            String moduleZipFile = FileHelper.getFileFromHash(fileTemp, hash);
+            String zipContainer = FileHelper.unzip(fileTemp + moduleZipFile, fileTemp);
+            if (!FileHelper.exists(zipContainer, "manifest.json")) {
+                throw new ConfException(StatusResponseType.DATA_MODULE_VERSION_INVALID_ARCHIVE.text(), StatusResponseType.DATA_MODULE_VERSION_INVALID_ARCHIVE.code());
+            }
+
+            //delete directory for zip extracting
+            FileUtils.deleteDirectory(new File(zipContainer));
+            //copy final file to repository
+            FileHelper.copyFromTempToRepo(fileTemp, fileRepository, hash);
+
             //save after all exceptions
             ModuleVersion moduleVersion = new ModuleVersion();
             moduleVersion.setModuleId(moduleId);
@@ -655,16 +669,31 @@ public class DataController implements DataContextUrl, PagesContextUrl {
         try {
             //do NOT check for empty file, but check for unique hash, upon file uploading
             String newHash = "";
+            String fileName = "";
             if (!uploadFile.isEmpty()) {
                 String oldHash = moduleVersionRepository.findOne(moduleVersionId).getHash();
                 newHash = FileHelper.saveUploadedFile(fileTemp, fileRepository, uploadFile, digestAlgorithm);
                 if (moduleVersionRepository.findByHash(newHash) != null) {
                     throw new ConfException(StatusResponseType.DATA_MODULE_VERSION_HASH_EXISTS.text(), StatusResponseType.DATA_MODULE_VERSION_HASH_EXISTS.code());
                 }
+
+                //check for manifest.json within archive
+                String moduleZipFile = FileHelper.getFileFromHash(fileTemp, newHash);
+                String zipContainer = FileHelper.unzip(fileTemp + moduleZipFile, fileTemp);
+                if (!FileHelper.exists(zipContainer, "manifest.json")) {
+                    throw new ConfException(StatusResponseType.DATA_MODULE_VERSION_INVALID_ARCHIVE.text(), StatusResponseType.DATA_MODULE_VERSION_INVALID_ARCHIVE.code());
+                }
+
+                //copy final file to repository
+                FileHelper.copyFromTempToRepo(fileTemp, fileRepository, newHash);
+                //delete directory for zip extracting
+                FileUtils.deleteDirectory(new File(zipContainer));
                 //find filename from hash and remove old file from file repository
-                String fileName = FileHelper.getFileFromHash(fileRepository, oldHash);
+                fileName = FileHelper.getFileFromHash(fileRepository, oldHash);
                 FileHelper.removeFile(fileRepository, fileName);
             }
+
+
 
             //proceed with persisting
             ModuleVersion moduleVersion = moduleVersionRepository.findOne(moduleVersionId);
